cdk/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==20.3.0
    # via
    #   cattrs
    #   jsii
aws-cdk.assets==1.94.1
    # via
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-s3-assets
aws-cdk.aws-applicationautoscaling==1.94.1
    # via aws-cdk.aws-lambda
aws-cdk.aws-autoscaling-common==1.94.1
    # via aws-cdk.aws-applicationautoscaling
aws-cdk.aws-cloudwatch==1.94.1
    # via
    #   -r requirements.in
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-sqs
aws-cdk.aws-codeguruprofiler==1.94.1
    # via aws-cdk.aws-lambda
aws-cdk.aws-ec2==1.94.1
    # via
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-lambda
aws-cdk.aws-ecr-assets==1.94.1
    # via aws-cdk.aws-lambda
aws-cdk.aws-ecr==1.94.1
    # via
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-lambda
aws-cdk.aws-efs==1.94.1
    # via aws-cdk.aws-lambda
aws-cdk.aws-events==1.94.1
    # via
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-s3
aws-cdk.aws-iam==1.94.1
    # via
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-autoscaling-common
    #   aws-cdk.aws-cloudwatch
    #   aws-cdk.aws-codeguruprofiler
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-events
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
aws-cdk.aws-kms==1.94.1
    # via
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
aws-cdk.aws-lambda==1.94.1
    # via -r requirements.in
aws-cdk.aws-logs==1.94.1
    # via
    #   -r requirements.in
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-lambda
aws-cdk.aws-s3-assets==1.94.1
    # via
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
aws-cdk.aws-s3==1.94.1
    # via
    #   -r requirements.in
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-s3-assets
aws-cdk.aws-signer==1.94.1
    # via aws-cdk.aws-lambda
aws-cdk.aws-sqs==1.94.1
    # via aws-cdk.aws-lambda
aws-cdk.aws-ssm==1.94.1
    # via aws-cdk.aws-ec2
aws-cdk.cloud-assembly-schema==1.94.1
    # via
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-ssm
    #   aws-cdk.core
    #   aws-cdk.cx-api
aws-cdk.core==1.94.1
    # via
    #   -r requirements.in
    #   aws-cdk.assets
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-autoscaling-common
    #   aws-cdk.aws-cloudwatch
    #   aws-cdk.aws-codeguruprofiler
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-signer
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
aws-cdk.cx-api==1.94.1
    # via
    #   aws-cdk.assets
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.core
aws-cdk.region-info==1.94.1
    # via
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-iam
    #   aws-cdk.core
boto3==1.17.33
    # via -r requirements.in
botocore==1.20.33
    # via
    #   boto3
    #   s3transfer
cattrs==1.3.0
    # via jsii
constructs==3.3.68
    # via
    #   aws-cdk.assets
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-autoscaling-common
    #   aws-cdk.aws-cloudwatch
    #   aws-cdk.aws-codeguruprofiler
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-signer
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
    #   aws-cdk.core
invoke==1.5.0
    # via -r requirements.in
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsii==1.25.0
    # via
    #   aws-cdk.assets
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-autoscaling-common
    #   aws-cdk.aws-cloudwatch
    #   aws-cdk.aws-codeguruprofiler
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-signer
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
    #   aws-cdk.cloud-assembly-schema
    #   aws-cdk.core
    #   aws-cdk.cx-api
    #   aws-cdk.region-info
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk.assets
    #   aws-cdk.aws-applicationautoscaling
    #   aws-cdk.aws-autoscaling-common
    #   aws-cdk.aws-cloudwatch
    #   aws-cdk.aws-codeguruprofiler
    #   aws-cdk.aws-ec2
    #   aws-cdk.aws-ecr
    #   aws-cdk.aws-ecr-assets
    #   aws-cdk.aws-efs
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-lambda
    #   aws-cdk.aws-logs
    #   aws-cdk.aws-s3
    #   aws-cdk.aws-s3-assets
    #   aws-cdk.aws-signer
    #   aws-cdk.aws-sqs
    #   aws-cdk.aws-ssm
    #   aws-cdk.cloud-assembly-schema
    #   aws-cdk.core
    #   aws-cdk.cx-api
    #   aws-cdk.region-info
    #   constructs
python-dateutil==2.8.1
    # via
    #   botocore
    #   jsii
python-dotenv==0.15.0
    # via -r requirements.in
s3transfer==0.3.6
    # via boto3
six==1.15.0
    # via python-dateutil
typing-extensions==3.7.4.3
    # via jsii
urllib3==1.26.4
    # via botocore
